Display memo categories and make description text expandable
The categories are shown as Chips in the upper right corner of the Card.

If the memo contains a description, the Card header acts as an expander for it.
In search results, the description is expanded by default and can not be hidden.
The actual link is displayed, but is pointed to internal /link/_id URL.